+++ src/main/java/kr/hhplus/be/server/infrastructure/order/OrderDetailJpaRepository.java
@@ -1,4 +1,4 @@
-package kr.hhplus.be.server.infrastructure.point.order;
+package kr.hhplus.be.server.infrastructure.order;
 
 import kr.hhplus.be.server.domain.order.entity.OrderDetail;
 import org.springframework.data.jpa.repository.JpaRepository;
