+++ src/main/java/kr/hhplus/be/server/infrastructure/order/OrderJpaRepository.java
@@ -1,0 +1,7 @@
+package kr.hhplus.be.server.infrastructure.order;
+
+import kr.hhplus.be.server.domain.order.entity.Order;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface OrderJpaRepository extends JpaRepository<Order, Long>, OrderCustomRepository {
+}
