--- src/main/java/kr/hhplus/be/server/infrastructure/point/order/OrderJpaRepository.java
@@ -1,7 +1,0 @@
-package kr.hhplus.be.server.infrastructure.point.order;
-
-import kr.hhplus.be.server.domain.order.entity.Order;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface OrderJpaRepository extends JpaRepository<Order, Long> {
-}
